# src/main/resources/openapi/sboot-ecobint-api.yaml
openapi: 3.1.0
info:
  title: Swagger Ecobint - OpenAPI 3.1
  version: 1.0.0
  description: API responsável pela comunicação entre dispositivos IoT de lixeiras inteligentes e a aplicação mobile da plataforma Ecobint. A API permite o envio do nível de enchimento das lixeiras para monitoramento, gerenciamento e acompanhamento em tempo real.

paths:
  /bin:
    post:
      tags:
        - Dispositivo
      summary: Adicionar nova lixeira.
      description: Endpoint utilizado pela aplicação mobile para adicionar uma nova lixeira.
      operationId: postNewBin
      requestBody:
        description: Informações da nova lixeira.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewBinRequest'
        required: true
      responses:
        '201':
          description: Lixeira criada com sucesso.
          headers:
            Location:
              description: URL para a nova lixeira criada.
              schema:
                type: string
                format: uri
        '400':
          description: Dados inválidos enviados na requisição.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '500':
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
  /bin/{binId}:
    get:
      tags:
        - Dispositivo
      summary: Obter informações da lixeira.
      description: Endpoint utilizado pela aplicação mobile para obter informações da lixeira.
      operationId: getBin
      parameters:
        - name: binId
          in: path
          description: ID da lixeira.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Informações da lixeira obtidas com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BinResponse'
        '400':
          description: Dados inválidos enviados na requisição.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '404':
          description: Lixeira não encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '500':
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
    post:
      tags:
        - Dispositivo
      summary: Enviar informações de enchimento da lixeira.
      description: Endpoint utilizado pelo dispositivo IoT para enviar o nível atual de enchimento da lixeira inteligente.
      operationId: postBinStatus
      parameters:
        - name: binId
          in: path
          description: ID da lixeira.
          required: true
          schema:
            type: string
      requestBody:
        description: Porcentagem da lixeira atual.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BinStatusRequest'
        required: true
      responses:
        '204':
          description: Status recebido com sucesso.
        '400':
          description: Dados inválidos enviados na requisição.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '500':
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
    put:
      tags:
        - Dispositivo
      summary: Atualizar informações da lixeira.
      description: Endpoint utilizado pelo dispositivo IoT para atualizar informações da lixeira.
      operationId: updateBin
      parameters:
        - name: binId
          in: path
          description: ID da lixeira.
          required: true
          schema:
            type: string
      requestBody:
        description: Informações atualizadas da lixeira.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBinRequest'
        required: true
      responses:
        '200':
          description: Informações da lixeira atualizadas com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BinResponse'
        '400':
          description: Dados inválidos enviados na requisição.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '500':
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
  /bins:
    get:
      tags:
        - Dispositivo
      summary: Obter informações de todas as lixeiras (provisório).
      description: Endpoint utilizado pela aplicação mobile para obter informações de todas as lixeiras cadastradas.
      operationId: getAllBins
      responses:
        '200':
          description: Informações das lixeiras obtidas com sucesso.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BinResponse'
        '400':
          description: Dados inválidos enviados na requisição.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
        '204':
          description: Nenhuma lixeira cadastrada.
        '500':
          description: Erro interno do servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
components:
  schemas:
    StandardError:
      type: object
      properties:
        codigo:
          type: integer
        mensagem:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - codigo
        - mensagem
        - timestamp
    BinStatusRequest:
      type: object
      properties:
        porcentagemEnchimento:
          type: number
          format: double
          minimum: 0
          maximum: 1
          example: 0.5
          description: Percentual atual de enchimento da lixeira.
      required:
        - porcentagemEnchimento
    BinResponse:
      type: object
      properties:
        id:
          type: string
        nome:
          type: string
        localizacao:
          type: string
        nivelEnchimento:
          type: number
          format: double
        ultimaAtualizacao:
          type: string
          format: date-time
    NewBinRequest:
      type: object
      properties:
        nome:
          type: string
          minLength: 1
          maxLength: 100
        localizacao:
          type: string
          minLength: 1
          maxLength: 100
      required:
        - nome
        - localizacao
    UpdateBinRequest:
      type: object
      properties:
        nome:
          type: string
          minLength: 1
          maxLength: 100
        localizacao:
          type: string
          minLength: 1
          maxLength: 100